Updated Logo and Main Menu sprites.
I added a pixelated version of the piano picture to make it more coherent with the in-game graphics, but if we want to revert back to the hi-res image I only have to switch off a layer in the aseprite file. Maybe the menu lettering should also be pixels, but what the hell. Maybe this is good enough?